1. Refactor code 2. Updated to typescript 3.9.2
