+++ src/collector/dashboard/collectDashboard.ts
