+++ src/collector/dashboard/index.ts
@@ -1,4 +1,4 @@
-export { collectDashboard } from './cacheDashboard'
+export { collectDashboard } from './collectDashboard'
 export { getAccountCountByDay } from './accountGrowth'
 export { getBlockRewardsByDay } from './blockReward'
 export { getTxVolumeByDay } from './txVolume'
